refactor(autocomplete): Refactor replace word at cursor and search by word logic
AutoComplete widget no longer depends on the DOM input element to determine the cursor position. All
logic which depends on the cursor is moved in the SearchBar component. SearchBar sends the filter
word (trimmed, without separator characters) to the AutoComplete component and also notifies when
the input text is changed by the user. The AutoComplete component holds two state values - the text
of the SearchBar and array of values. AutoComplete updates the text state when the user types in the
input or when the user selects item from the List. Text state is not updated during search!