fix(dropdownlist): component's event handling
- onClose and onOpen callbacks of the stateless component are merged into
one onToggle that receives true/false argument which indicates popup
visibility state.

- stateless component call onChange when it is blurred. The parent
component which holds the state closes it through the props.

- when popup is opened the stateless component adds mousedown event listener to
the documentElement in order to detect clicks outside of the component
area. If such click is detected the widget fires its change event, the
parent stateful component closes it.

- the aforementioned event listener is removed when the popup closes or
when the component is about to unmount