fix(dropdownlist): filtering behaviour
Issues:

1. The filter input randomly clears if the user opens/closes the popup
rapidly.
2. There is no suitable callback that will allow the user to clear the
filter and reset data after item is selected through the keyboard.

Solution:

- ListFilter was changed from uncontrolled to controlled input. Its
value is stored in the DropDownList state.filter and is updated every
time the `onFilter` callback is about to be invoked.

- As a result of this the `delay` configuration property will not have
effect over the filtering behaviour. The `delay` will be honored only
for the internal DropDownList search functionality. In case of filtering
the high order component can decide how often to update the DropDownList data according to the filter text.

- The `onFilter` callback will be invoked with empty string as an
argument after DropDownList `onChange` is called. In this way the
developer will be able to clear the data filter after user selects item
from the list either via the mouse or by pressing the "Enter" key.

- The docs and tests are updated accordingly